% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{s3_auth}
\alias{s3_auth}
\title{Request Amazon S3 credentials}
\usage{
s3_auth(edl_token = "", user = "", pass = "", verbose = FALSE)
}
\arguments{
\item{edl_token}{EDL token value as string. Optional if already store in env.}

\item{user}{Username string, alternative to token when used with password.}
}
\value{
pass Password string, alternative to token when used with username.

verbose Boolean to flag output verbosity.
}
\description{
Gets a AWS key and token for NISAR S3 bucket.
}
